Rename _interfaces to _i_observer
